Update to work with jQuery 1.9
In jQuery 1.9, the `.live` function has been removed and the `.selector` property has been deprecated.
We’re now using `.on` internally, replacing both `.live` as well as `.bind`. This means that jQuery 1.7 is required.

Up to and including jQuery 1.9, the live setting is still supported and works as it did before.
The recommended way to use it, however, has changed to

    $(document).popover({live: '.selector'})

which works in all versions of jQuery starting with 1.7 up to 2.0 and possibly beyond.